refactor base tagdb into its own file
