+++ webapp/graphite/tags/localdatabase.py
@@ -3,7 +3,8 @@ import re
 from django.db import connection
 from hashlib import sha256
 
-from graphite.tags.utils import BaseTagDB, TaggedSeries
+from graphite.tags.base import BaseTagDB, TaggedSeries
+
 
 class LocalDatabaseTagDB(BaseTagDB):
   def find_series_query(self, tags):
